requirements.txt
python-dotenv
pandas
numpy
pymongo==4.5.0
certifi
pymongo[srv]
scikit-learn
mlflow
pyaml
dagshub
## -e .